Format common_working_graph for dwave-system
It was originally written for the dwave_networkx project
