+++ dwave/system/__init__.py
@@ -12,7 +12,7 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-# ================================================================================================
+# =============================================================================
 import dwave.system.flux_bias_offsets
 
 from dwave.system.samplers import *
@@ -20,3 +20,5 @@ import dwave.system.samplers
 
 from dwave.system.composites import *
 import dwave.system.composites
+
+from dwave.system.utilities import *
